// types/student.ts
export interface Student {
  id: number
  nombres: string
  apellidos: string
  codigo_estudiante?: string
  fecha_nacimiento?: string
  genero?: string
  direccion?: string
  telefono?: string
  correo_electronico?: string
  grado?: string
  seccion?: string
  estado?: string
  url_foto_perfil?: string
  createdAt?: string
  updatedAt?: string
  // Agrega más campos según sea necesario
}